+++ libs/asset-editor/src/lib/components/asset-editor-tabs/asset-editor-general/asset-editor-general.component.ts
@@ -1,9 +1,10 @@
 import { Component, inject, Input, OnDestroy, OnInit } from '@angular/core';
-import { fromAppShared, TranslatedValue } from '@asset-sg/client-shared';
+import { fromAppShared } from '@asset-sg/client-shared';
 import { AssetEditDetail, ValueItem } from '@asset-sg/shared';
 import { Role, SimpleWorkgroup } from '@asset-sg/shared/v2';
 import { Store } from '@ngrx/store';
 import { combineLatestWith, map, Observable, startWith, Subscription } from 'rxjs';
+import { TranslatedValueItem } from '../../../models/translated-value-item.interface';
 import { AssetForm } from '../../asset-editor-page/asset-editor-page.component';
 
 @Component({
@@ -96,11 +97,6 @@ export class AssetEditorGeneralComponent implements OnInit, OnDestroy {
     ]);
     this.form.markAsDirty();
   }
-}
-
-export interface TranslatedValueItem {
-  code: string;
-  value: TranslatedValue;
 }
 
 export const mapValueItemsToTranslatedItem = (item: Record<string, ValueItem> | null): TranslatedValueItem[] => {
