+++ libs/asset-editor/src/lib/models/translated-value-item.interface.ts
@@ -1,0 +1,6 @@
+import { TranslatedValue } from '@asset-sg/client-shared';
+
+export interface TranslatedValueItem {
+  code: string;
+  value: TranslatedValue;
+}
